# Fixes #90: Perform operations on the key of an object or array