The problem is that when we're iterating through, we lose track of the key of an array or key of an object, so any distinct op nodes on that key were never executed. 
The proposed solution address this specific scenario and leaves the others untouched (I believe).