@brett19 2019-05-03
Hey @nelio2k ,
Can you elaborate a bit more on the problem this is solving?  I was looking at the example code and what not, but am having trouble seeing what it enables that was not previously possible.
Cheers, Brett