@nelio2k 2019-05-03
This is more of a specific use case of doing an op on the key of an array or an object in a JSON, without referring to a specific JSON field underneath the key. 
Given the following JSON:
`{
  "objKey": {
    "key":"value"
  }
}`
Before the PR, the code does the following:

1. Sees the {
2. Goes into matchObjAndArr
3. Gets "objKey", stores it as `keyString`
4. Gets `:` and subsequently `{`
5. Execute the matchOp on node.Elem["objKey"], which recursively goes and perform on `key` and `value`.

Normally that's fine if users do ops on `key` within `objKey`, but if the "objKey" itself has ops, they don't get executed, and becomes lost as gojsonsm essentially "steps over" the "objKey" and the operations in its corresponding ExecNode.

The example that I found was if say an expression of `EXISTS(objKey)`.
Hope that makes sense?