@brett19 2019-05-10
A nil after token is acceptable, it is used to indicate that no after
blocks exist.  Nil is faster to detect than any other type of value.

Cheers, Brett

On Thu., May 9, 2019, 3:54 p.m. ysui6888, <notifications@github.com> wrote:

> *@ysui6888* commented on this pull request.
> ------------------------------
>
> In fastMatcher.go
> <https://github.com/couchbaselabs/gojsonsm/pull/105#discussion_r282695321>
> :
>
> > @@ -625,6 +626,33 @@ func (m *FastMatcher) matchExec(token tokenType, tokenData []byte, tokenDataLen
>  	return nil
>  }
>
> +// If any operation is done strictly on the key of an obj or array, this method takes care of it
> +func (m *FastMatcher) matchKeyOfObj(token tokenType, tokenData []byte, node *ExecNode) (error, bool) {
> +	lastArrOrMapKeyByteLen := len(m.lastArrOrMapKeyBytes)
> +	if lastArrOrMapKeyByteLen == 0 {
> +		return nil, false
> +	}
> +
> +	var savedKey []byte = m.lastArrOrMapKeyBytes
> +	m.lastArrOrMapKeyBytes = nil
> +	if len(node.Ops) == 0 || (token != tknObjectStart && token != tknArrayStart) {
> +		return nil, false
> +	}
> +
> +	// Since this matchExec should only execute on the specific key
> +	// temporarily don't do any afters, since they are out of scope
>
> i take it that we should not set node.after to nil, then?
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/couchbaselabs/gojsonsm/pull/105#discussion_r282695321>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AAML46ZJWSMCASAPVQTK5VTPUSTSBANCNFSM4HKOTNPA>
> .
>