@brett19 2019-05-11
I also don't like the idea of having the After temporarily rewritten.  It also convolutes the logic of processing, and causes jsonsm to need to iterate over a key numerous times.  We should be able to simply perform the Ops in the array and object branches of matchExec after its already processed it as an array/object.  This will also have the benefit of likely already knowing the bounds of the underlying JSON value for parsing, and in cases such as 'EXISTS', it doesn't even need to parse anyways.