@nelio2k 2019-05-13
I've updated it by removing the afternode modification, and added comparison on the Execnode of the object or array itself.