@nelio2k 2019-05-23
@brett19  I couldn't find your latest comment on this PR. I did forget to run the benchmark test.

With regards to your recommendation of checking for ops to process... interestingly though, when I added the following check:
+               if len(node.Ops) > 0 {
... create NewObjectFastVal/NewArrayFastVal

The performance actually and consistently got worse.

Current PR code without the len() check above:
```
neil.huang@NeilsMacbookPro:~/source/couchbase/godeps/src/github.com/couchbase/gojsonsm$ go test -bench=.
goos: darwin
goarch: amd64
pkg: github.com/couchbase/gojsonsm
BenchmarkBinTree-8             	20000000	        70.6 ns/op
BenchmarkParseString-8         	100000000	        10.8 ns/op
BenchmarkParseBigString-8      	100000000	        10.9 ns/op
BenchmarkParseEscString-8      	30000000	        52.5 ns/op
BenchmarkParseBigEscString-8   	10000000	       172 ns/op
BenchmarkParseInteger-8        	100000000	        12.3 ns/op
BenchmarkParseNumber-8         	30000000	        47.7 ns/op
BenchmarkParseNull-8           	200000000	         7.06 ns/op
BenchmarkParseTrue-8           	200000000	         7.01 ns/op
BenchmarkParseFalse-8          	200000000	         8.41 ns/op
BenchmarkTokenize-8            	   30000	     44433 ns/op	 351.67 MB/s
PASS
ok  	github.com/couchbase/gojsonsm	18.548s

```
With checking len(node.Ops) before creating Fastval:

```
neil.huang@NeilsMacbookPro:~/source/couchbase/godeps/src/github.com/couchbase/gojsonsm$ go test -bench=.
goos: darwin
goarch: amd64
pkg: github.com/couchbase/gojsonsm
BenchmarkBinTree-8             	20000000	        74.5 ns/op
BenchmarkParseString-8         	100000000	        11.5 ns/op
BenchmarkParseBigString-8      	100000000	        11.5 ns/op
BenchmarkParseEscString-8      	30000000	        55.3 ns/op
BenchmarkParseBigEscString-8   	10000000	       183 ns/op
BenchmarkParseInteger-8        	100000000	        13.7 ns/op
BenchmarkParseNumber-8         	30000000	        53.0 ns/op
BenchmarkParseNull-8           	200000000	         7.74 ns/op
BenchmarkParseTrue-8           	200000000	         7.77 ns/op
BenchmarkParseFalse-8          	200000000	         9.00 ns/op
BenchmarkTokenize-8            	   30000	     52628 ns/op	 296.91 MB/s
PASS
ok  	github.com/couchbase/gojsonsm	20.218s
```

I'm trying to think of reasons why... the only things I can think of are:
1. Creating a NewFastValObject is cheap ... we're really just creating slice references to existing memory and perhaps golang has got it optimized.
2. Perhaps the len() call really is as expensive as we have thought, and thus why we tried to remove the len() calls previously in gojsonsm and goxdcr.

Do you or @ysui6888  have any ideas on why as is it's better performant? 
Anything else that concerns you?