@brett19 2019-05-23
Why not try checking if `node.Ops == nil`?